refactor(modeling) moved measurements out of operations
