+++ packages/modeling/src/index.js
@@ -3,6 +3,7 @@ module.exports = {
   connectors: require('./connectors'),
   geometries: require('./geometries'),
   maths: require('./maths'),
+  measurements: require('./measurements'),
   primitives: require('./primitives'),
   text: require('./text'),
   utils: require('./utils'),
@@ -11,6 +12,5 @@ module.exports = {
   expansions: require('./operations/expansions'),
   extrusions: require('./operations/extrusions'),
   hulls: require('./operations/hulls'),
-  measurements: require('./operations/measurements'),
   transforms: require('./operations/transforms')
 }
