+++ packages/modeling/src/measurements/measureVolume.js
@@ -1,9 +1,9 @@
-const flatten = require('../../utils/flatten')
+const flatten = require('../utils/flatten')
 
-const geom2 = require('../../geometries/geom2')
-const geom3 = require('../../geometries/geom3')
-const path2 = require('../../geometries/path2')
-const poly3 = require('../../geometries/poly3')
+const geom2 = require('../geometries/geom2')
+const geom3 = require('../geometries/geom3')
+const path2 = require('../geometries/path2')
+const poly3 = require('../geometries/poly3')
 
 /*
  * Measure the volume of the given geometry.
